Sounds!
Grabbed some random placeholder sounds from the internet.
Sounds are sort of working: HTML5 is kind of screwy with audio, the
system isn't very well flushed out. Because of this, I've refactored to
put the minimum stress on it possible. This may need to be tested on
multiple browsers, as not all of them fully support HTML5 audio
features.